Improved LoadStoryTask and initial load
Fixed LoadStoryTask so that the activity is no longer referenced in
doInBackground (so we aren't actually holding onto the reference unless
we're actually doing something involving the activity).  Also removed
the explicit story file path in LoadStoryTask and replaced it with a
getter to the file path in the activity.  Lastly, added more meaningful
default values for text views.